requirements.txt
SimpleAES==1.0
benchmark==0.1.5
pycrypto==2.6
